// src/generated/models/Cr22e_excepcionescalidaddatosesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_excepcionescalidaddatosescr22e_ambito = {
  100000001: 'Global',
  100000002: 'Sucursal'
} as const;
export type Cr22e_excepcionescalidaddatosescr22e_ambito = keyof typeof Cr22e_excepcionescalidaddatosescr22e_ambito;
export const Cr22e_excepcionescalidaddatosescr22e_campoevaluado = {
  100000001: 'DireccionIP',
  100000002: 'Hostname',
  100000003: 'Numerodeserie',
  100000004: 'ClaveAnydesk',
  100000005: 'Sucursal',
  100000006: 'Departamento',
  100000007: 'Ubicacion',
  100000008: 'Responsable',
  100000009: 'Estado',
  100000010: 'Otro'
} as const;
export type Cr22e_excepcionescalidaddatosescr22e_campoevaluado = keyof typeof Cr22e_excepcionescalidaddatosescr22e_campoevaluado;
export const Cr22e_excepcionescalidaddatosescr22e_tipoproblema = {
  100000001: 'Duplicado',
  100000002: 'Datofaltante',
  100000003: 'Valorsospechoso',
  100000004: 'Inconsistencia'
} as const;
export type Cr22e_excepcionescalidaddatosescr22e_tipoproblema = keyof typeof Cr22e_excepcionescalidaddatosescr22e_tipoproblema;
export const Cr22e_excepcionescalidaddatosesstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_excepcionescalidaddatosesstatecode = keyof typeof Cr22e_excepcionescalidaddatosesstatecode;
export const Cr22e_excepcionescalidaddatosesstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_excepcionescalidaddatosesstatuscode = keyof typeof Cr22e_excepcionescalidaddatosesstatuscode;

export interface Cr22e_excepcionescalidaddatosesBase {
  cr22e_activa?: boolean;
  cr22e_ambito?: Cr22e_excepcionescalidaddatosescr22e_ambito;
  cr22e_campoevaluado?: Cr22e_excepcionescalidaddatosescr22e_campoevaluado;
  cr22e_clavealerta?: string;
  cr22e_equiposafectados?: string;
  cr22e_excepcionescalidaddatosid: string;
  cr22e_fechaignorado?: string;
  cr22e_fechareactivacion?: string;
  cr22e_motivo?: string;
  cr22e_name: string;
  "cr22e_Sucursales@odata.bind"?: string;
  cr22e_tipoproblema?: Cr22e_excepcionescalidaddatosescr22e_tipoproblema;
  cr22e_valordetectado?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_excepcionescalidaddatosesstatecode;
  statuscode?: Cr22e_excepcionescalidaddatosesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_excepcionescalidaddatoses extends Cr22e_excepcionescalidaddatosesBase {
  cr22e_activaname?: string;
  cr22e_ambitoname?: string;
  cr22e_campoevaluadoname?: string;
  cr22e_sucursalesname?: string;
  cr22e_tipoproblemaname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr22e_sucursales?: object;
  _cr22e_sucursales_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}